refactor(desktop): extract the onKey keyboard handler into keydown modules (F5.8)
The ~710-line global keydown handler moves out of App.tsx, split to stay
under 500 lines each:

- keydownCtx.ts (151): the KeydownCtx type (~130 values/handlers/refs the
  handler needs) + VimOp
- keydownHandler.ts (299): the prefix — zoom, advanced-search, Ctrl/Cmd
  guard, help, the layered-Escape modal chain, action-plan shortcuts, the
  typing/drafts guards, accounts toggle, content-search n/N
- keydownMain.ts (465): reader scrolling, j/k/arrow navigation, and the
  chord -> action lookup incl. VIM range ops

App's keydown useEffect now builds the ctx and wires handleKeyDown; the
handler bodies are byte-identical (behavior preserved by construction).

App.tsx 2,574 -> 1,888. New keyboard.spec.ts covers c/compose, j/k nav,
?/help + Escape (previously uncovered). unit 61 / e2e 39 green, tsc +
build clean.
