+++ desktop/frontend/src/App.tsx
@@ -20,12 +20,10 @@ import {
 } from "./api";
 import type { ComposeInit } from "./Compose";
 import {
-  matchesCombo,
   emailAddr,
   cleanSubject,
-  countMatches,
 } from "./format";
-import { replyInit, replyAllInit, forwardInit } from "./compose";
+import { replyInit, forwardInit } from "./compose";
 import { activeAiPanel } from "./aiPanels";
 import ModalsPrimary from "./ModalsPrimary";
 import ModalsSecondary from "./ModalsSecondary";
@@ -41,6 +39,8 @@ import { useMailActions } from "./useMailActions";
 import { type AdvFilters, EMPTY_ADV } from "./advancedSearch";
 import { runCommand } from "./commandRunner";
 import { useActionPlan } from "./useActionPlan";
+import { handleKeyDown } from "./keydownHandler";
+import type { KeydownCtx } from "./keydownCtx";
 import { useBootstrap } from "./useBootstrap";
 import { useZoom } from "./useZoom";
 import { useTheme } from "./useTheme";
@@ -1307,718 +1307,32 @@ export default function App() {
   // cursor without opening; Enter opens. This avoids marking mail read while
   // just scanning.
   useEffect(() => {
-    const onKey = (e: KeyboardEvent) => {
-      const tag = (e.target as HTMLElement | null)?.tagName;
-      const typing = tag === "INPUT" || tag === "TEXTAREA";
-      const chord = e.key === " " ? "space" : e.key;
-
-      // UI zoom (Cmd/Ctrl +/-/0) — handled first so it works everywhere, even
-      // over modals or while typing. WKWebView ignores native zoom.
-      if ((e.metaKey || e.ctrlKey) && !e.altKey) {
-        if (e.key === "=" || e.key === "+") {
-          e.preventDefault();
-          bumpZoom(0.1);
-          return;
-        }
-        if (e.key === "-" || e.key === "_") {
-          e.preventDefault();
-          bumpZoom(-0.1);
-          return;
-        }
-        if (e.key === "0") {
-          e.preventDefault();
-          resetZoom();
-          return;
-        }
-      }
-
-      // Advanced search builder (default Ctrl+F / Cmd+F, from search_advanced).
-      // Handled before the generic Ctrl/Cmd early-return and the typing guard so
-      // it opens from anywhere — list, reader, or the search box — like the TUI.
-      // Only modifier combos are honored here (a bare key would hijack typing).
-      if (keymap.searchAdvanced && !advOpen && matchesCombo(e, keymap.searchAdvanced)) {
-        e.preventDefault();
-        setAdvOpen(true);
-        return;
-      }
-
-      // Never let OS/browser clipboard & navigation combos (Cmd/Ctrl+C, V, X, A,
-      // Z…) fall through to single-key actions like compose ("c"). The only
-      // Cmd/Ctrl combos we act on are zoom (handled above) and the accounts
-      // switcher (Cmd/Ctrl+A, handled below), so let everything else reach the
-      // browser — otherwise selecting text and pressing Cmd+C opened the composer.
-      if (e.metaKey || e.ctrlKey) {
-        const isAccounts =
-          (e.key === "a" || e.key === "A") && accounts.length > 1;
-        if (!isAccounts) return;
-      }
-
-      if (showHelp) {
-        if (e.key === "Escape" || chord === keymap.help) {
-          setShowHelp(false);
-          e.preventDefault();
-        }
-        return;
-      }
-      const anyModal =
-        compose ||
-        labelsFor ||
-        bulkLabels ||
-        promptsOpen ||
-        promptManagerOpen ||
-        linksFor ||
-        suggestFor ||
-        cmdOpen ||
-        queriesOpen ||
-        saveQueryOpen ||
-        planOpen ||
-        themePickerOpen ||
-        rulesOpen ||
-        promptPreview !== null ||
-        advOpen ||
-        statsOpen ||
-        configOpen ||
-        moveFor ||
-        bulkMove ||
-        rsvpPickerOpen ||
-        detRulesOpen ||
-        accountsOpen ||
-        attachmentsOpen ||
-        bulkPromptText !== null;
-      if (anyModal) {
-        // Escape closes the topmost modal from the window (WKWebView won't focus
-        // a bare div, so per-modal Escape handlers on divs are unreliable). Order
-        // = last-opened first, so a sub-modal (e.g. rules over the plan) closes
-        // before its parent. Pickers also self-close via their own listener;
-        // double-closing is harmless.
-        if (e.key === "Escape") {
-          e.preventDefault();
-          if (accountsOpen) setAccountsOpen(false);
-          else if (attachmentsOpen) setAttachmentsOpen(false);
-          else if (promptPreview !== null) setPromptPreview(null);
-          else if (rulesOpen) setRulesOpen(false);
-          else if (bulkPromptText !== null) setBulkPromptText(null);
-          else if (saveQueryOpen) setSaveQueryOpen(false);
-          else if (moveFor) setMoveFor(null);
-          else if (bulkMove) setBulkMove(false);
-          else if (suggestFor) setSuggestFor(null);
-          else if (advOpen) setAdvOpen(false);
-          else if (statsOpen) setStatsOpen(false);
-          else if (configOpen) setConfigOpen(false);
-          else if (planPreview) setPlanPreview(null);
-          else if (planMove) setPlanMove(null);
-          else if (planOpen) setPlanOpen(false);
-          else if (themePickerOpen) setThemePickerOpen(false);
-          else if (queriesOpen) setQueriesOpen(false);
-          else if (rsvpPickerOpen) setRsvpPickerOpen(false);
-          else if (linksFor) setLinksFor(null);
-          else if (bulkLabels) setBulkLabels(false);
-          else if (labelsFor) setLabelsFor(null);
-          else if (promptsOpen) setPromptsOpen(false);
-          else if (promptManagerOpen) setPromptManagerOpen(false);
-          else if (cmdOpen) setCmdOpen(false);
-          else if (compose) setCompose(null);
-          return;
-        }
-        // Ctrl/Cmd+A toggles the account menu closed too (it's what opened it),
-        // since the menu is now part of the modal guard and swallows other keys.
-        if (
-          accountsOpen &&
-          (e.ctrlKey || e.metaKey) &&
-          (e.key === "a" || e.key === "A")
-        ) {
-          e.preventDefault();
-          setAccountsOpen(false);
-          return;
-        }
-        // Action-plan reachable-by-keyboard shortcuts for its header buttons.
-        if (
-          planOpen &&
-          !rulesOpen &&
-          !detRulesOpen &&
-          promptPreview === null &&
-          planMove === null &&
-          planPreview === null &&
-          bulkPromptText === null
-        ) {
-          if (e.key === "r") {
-            e.preventDefault();
-            if (rulesEnabled) void openRules();
-            return;
-          }
-          if (e.key === "p") {
-            e.preventDefault();
-            void viewAnalyzerPrompt();
-            return;
-          }
-          // l applies a label bucket as LABEL-ONLY (Enter does the move variant).
-          if (e.key === "l") {
-            const node = planNodesRef.current[planActiveRef.current];
-            const c = node ? plan?.categories[node.catIdx] : undefined;
-            if (c && c.action === "label") {
-              e.preventDefault();
-              void applyCategory(c, false);
-              return;
-            }
-          }
-          // m reassigns to another bucket: on an email node, that one email; on a
-          // category node, the whole category (the TUI's action-plan move).
-          if (e.key === "m") {
-            const node = planNodesRef.current[planActiveRef.current];
-            if (node) {
-              e.preventDefault();
-              if (node.type === "email") {
-                setPlanMove({
-                  kind: "email",
-                  catIdx: node.catIdx,
-                  id: node.id,
-                });
-              } else {
-                setPlanMove({ kind: "category", catIdx: node.catIdx });
-              }
-              return;
-            }
-          }
-          // Space toggles selection of the active email (deselect to exclude it
-          // from the category's apply/move); on a category it expands/collapses.
-          if (e.key === " ") {
-            const node = planNodesRef.current[planActiveRef.current];
-            if (node?.type === "email") {
-              e.preventDefault();
-              setPlanExcluded((prev) => {
-                const n = new Set(prev);
-                if (n.has(node.id)) n.delete(node.id);
-                else n.add(node.id);
-                return n;
-              });
-              return;
-            }
-            const cat = node ? plan?.categories[node.catIdx] : undefined;
-            if (cat) {
-              e.preventDefault();
-              setExpandedCats((prev) => {
-                const n = new Set(prev);
-                if (n.has(cat.name)) n.delete(cat.name);
-                else n.add(cat.name);
-                return n;
-              });
-              return;
-            }
-          }
-          // → expands the active category (or the parent of the active email);
-          // ← collapses it.
-          if (e.key === "ArrowRight" || e.key === "ArrowLeft") {
-            const node = planNodesRef.current[planActiveRef.current];
-            const cat = node ? plan?.categories[node.catIdx] : undefined;
-            if (cat) {
-              e.preventDefault();
-              setExpandedCats((prev) => {
-                const n = new Set(prev);
-                if (e.key === "ArrowLeft") n.delete(cat.name);
-                else n.add(cat.name);
-                return n;
-              });
-              return;
-            }
-          }
-        }
-        return;
-      }
-      if (typing) {
-        if (e.key === "Escape") {
-          // preventDefault so macOS/WKWebView doesn't treat Escape as
-          // "leave fullscreen"; we handle it ourselves.
-          e.preventDefault();
-          const el = e.target as HTMLElement;
-          if (el === searchRef.current) {
-            // TUI parity: Escape in the search box clears the filter and
-            // returns to the default inbox, instead of just blurring.
-            setQuery("");
-            if (localFilter) {
-              setLocalFilter(false);
-              setMessages(fullMessagesRef.current);
-            } else if (activeQuery) {
-              void load("");
-            }
-          }
-          el.blur();
-        }
-        return;
-      }
-      if (draftsView) {
-        if (e.key === "Escape" || chord === keymap.drafts) {
-          setDraftsView(false);
-          e.preventDefault();
-        } else if (chord === keymap.help) {
-          setShowHelp(true);
-        }
-        return;
-      }
-
-      // Ctrl/Cmd+A opens the account switcher (TUI's accounts shortcut). Placed
-      // after the typing guard so it never hijacks select-all inside inputs.
-      if (
-        (e.ctrlKey || e.metaKey) &&
-        (e.key === "a" || e.key === "A") &&
-        accounts.length > 1
-      ) {
-        e.preventDefault();
-        setAccountsOpen((v) => !v);
-        return;
-      }
-
-      // Content-search match navigation: once a find is active, n = next match,
-      // N = previous (the TUI's in-message n/N). Typing in the search box is
-      // handled by the typing guard above, so this only fires from the reader.
-      if (csOpen && csQuery && detail && (chord === "n" || chord === "N")) {
-        e.preventDefault();
-        const total = countMatches(detail.plainText || "", csQuery);
-        if (total > 0)
-          setCsIndex((i) =>
-            chord === "n" ? (i + 1) % total : (i - 1 + total) % total,
-          );
-        return;
-      }
-
-      const idx = selectedId
-        ? messages.findIndex((m) => m.id === selectedId)
-        : -1;
-      // Navigation previews the message (shows content, scrolls to it) without
-      // marking it read — in bulk mode too, so the reader follows the cursor
-      // like the TUI (previewMessage also sets selectedId as the highlight).
-      const moveCursor = (i: number) => {
-        if (i < 0 || i >= messages.length) return;
-        previewMessage(messages[i]);
-      };
-      const hasSel = bulkMode && selected.size > 0;
-
-      // Layered Escape: an open AI panel (summary / prompt / reformat) is closed
-      // first, before Escape hands the reader back / exits a search.
-      if (chord === "Escape" && dismissAI()) {
-        e.preventDefault();
-        return;
-      }
-
-      // --- reader-focused scrolling (TUI parity) ---
-      // After Enter/click-open (or a click in the reader), arrows & j/k scroll
-      // the message body instead of moving the inbox cursor; Space/PageDown page
-      // through it; Escape hands focus back to the list (a second Escape then
-      // closes the reader). Only nav keys are intercepted here — actions
-      // (archive, reply, summarize, …) still fall through to the switch below.
-      if (readerFocused && detail && !bulkMode) {
-        const sc = readerBodyRef.current;
-        const page = sc ? sc.clientHeight * 0.9 : 400;
-        if (chord === "j" || chord === "ArrowDown") {
-          e.preventDefault();
-          sc?.scrollBy({ top: 80 });
-          return;
-        }
-        if (chord === "k" || chord === "ArrowUp") {
-          e.preventDefault();
-          sc?.scrollBy({ top: -80 });
-          return;
-        }
-        if (chord === "PageDown") {
-          e.preventDefault();
-          sc?.scrollBy({ top: page });
-          return;
-        }
-        // NB: Space is intentionally NOT a reader-scroll key. It always means
-        // "select this message" (bulk select) — even while a message is open — so
-        // it falls through to the selection handler below. Page the reader with
-        // PageDown / j / ArrowDown instead.
-        if (chord === "PageUp") {
-          e.preventDefault();
-          sc?.scrollBy({ top: -page });
-          return;
-        }
-        if (chord === "Home") {
-          e.preventDefault();
-          sc?.scrollTo({ top: 0 });
-          return;
-        }
-        if (chord === "End") {
-          e.preventDefault();
-          sc?.scrollTo({ top: sc.scrollHeight });
-          return;
-        }
-        if (chord === "Enter") {
-          // Already open — don't reload/re-mark; just consume it.
-          e.preventDefault();
-          return;
-        }
-        if (chord === "Escape") {
-          e.preventDefault();
-          setReaderFocused(false);
-          return;
-        }
-      }
-
-      // --- list navigation (not remappable, like the TUI table) ---
-      if (chord === "j" || chord === "ArrowDown") {
-        e.preventDefault();
-        moveCursor(idx < 0 ? 0 : Math.min(messages.length - 1, idx + 1));
-        return;
-      }
-      if (chord === "k" || chord === "ArrowUp") {
-        e.preventDefault();
-        moveCursor(idx <= 0 ? 0 : idx - 1);
-        return;
-      }
-      if (chord === "Enter") {
-        e.preventDefault();
-        if (idx >= 0) {
-          if (bulkMode) toggleSelect(messages[idx].id);
-          else void openMessage(messages[idx]);
-        }
-        return;
-      }
-      // Bulk selection toggle. Handled here in the list-nav block — not via the
-      // remappable action switch — so it fires reliably on the highlighted row.
-      // Outside bulk mode it enters bulk mode and selects the current row;
-      // inside, it toggles and advances like the TUI.
-      // Space always works (the TUI default and what everyone reaches for), in
-      // addition to the configured bulk_select key (a literal " " also means
-      // space). Space isn't a remappable action here, so this never clashes.
-      // Exception: if bulk_select is bound to the SAME key as search (some
-      // configs put "s" on both), that key must NOT hijack search in the list —
-      // it falls through to the search action below, and bulk-select stays on
-      // Space. Otherwise the search key is silently shadowed and never searches.
-      const bulkKey =
-        keymap.bulkSelect === " " ? "space" : keymap.bulkSelect || "space";
-      const bulkKeyShadowsSearch =
-        bulkKey !== "space" && bulkKey === keymap.search;
-      // Space ALWAYS selects. The configured bulk key selects too — unless it
-      // collides with search, in which case that key searches instead (Space
-      // still selects, so bulk is never lost).
-      if (
-        (chord === "space" || (chord === bulkKey && !bulkKeyShadowsSearch)) &&
-        messages.length > 0
-      ) {
-        e.preventDefault();
-        const i = idx >= 0 ? idx : 0;
-        if (i < messages.length) {
-          // Toggle the highlighted message in place — never auto-advance
-          // (matches the TUI; you move with j/k). Entering bulk selects it.
-          if (!bulkMode) setBulkMode(true);
-          toggleSelect(messages[i].id);
-          setSelectedId(messages[i].id);
-        }
-        return;
-      }
-      if (chord === "Escape") {
-        // preventDefault so WKWebView doesn't leave fullscreen on macOS when we
-        // consume Escape to close the reader / exit a search (TUI parity).
-        if (bulkMode) {
-          e.preventDefault();
-          exitBulk();
-        } else if (detail) {
-          e.preventDefault();
-          setSelectedId(null);
-          setDetail(null);
-          setReaderFocused(false);
-        } else if (activeQuery || localFilter) {
-          // No reader/bulk open: Escape exits the active search back to the
-          // default inbox (TUI parity).
-          e.preventDefault();
-          setQuery("");
-          if (localFilter) {
-            setLocalFilter(false);
-            setMessages(fullMessagesRef.current);
-          } else {
-            void load("");
-          }
-        }
-        return;
-      }
-      if (chord === "*") {
-        if (bulkMode) {
-          e.preventDefault();
-          setSelected(new Set(messages.map((m) => m.id)));
-        }
-        return;
-      }
-      if (chord === keymap.contentSearch) {
-        e.preventDefault();
-        // With a message open, "/" searches within its body (like the TUI).
-        // Otherwise it focuses the inbox search box.
-        if (detail && !bulkMode) {
-          setViewHtml(false); // content search runs over the plain-text body
-          setCsOpen(true);
-          setCsQuery("");
-          setCsIndex(0);
-        } else {
-          searchRef.current?.focus();
-        }
-        return;
-      }
-
-      // --- goto top/bottom ---
-      if (chord === keymap.gotoBottom) {
-        e.preventDefault();
-        moveCursor(messages.length - 1);
-        return;
-      }
-      if (keymap.gotoTop === "gg") {
-        if (chord === "g") {
-          e.preventDefault();
-          const now = Date.now();
-          if (now - gPressedAt.current < (keymap.vimTimeoutMs || 1000)) {
-            moveCursor(0);
-            gPressedAt.current = 0;
-          } else {
-            gPressedAt.current = now;
-          }
-          return;
-        }
-      } else if (chord === keymap.gotoTop) {
-        e.preventDefault();
-        moveCursor(0);
-        return;
-      }
-
-      // --- VIM range operations (a3a, d2d, t5t, l2l) ---
-      // A single press of a range-op key fires the normal single-message action
-      // after vimRangeTimeoutMs; key-count-key applies it to the next N rows.
-      // This mirrors the TUI, where the same keys buffer before acting.
-      const rangeOps: Record<
-        string,
-        "archive" | "trash" | "toggleRead" | "manageLabels"
-      > = {};
-      if (keymap.archive) rangeOps[keymap.archive] = "archive";
-      if (keymap.trash) rangeOps[keymap.trash] = "trash";
-      if (keymap.toggleRead) rangeOps[keymap.toggleRead] = "toggleRead";
-      if (keymap.manageLabels) rangeOps[keymap.manageLabels] = "manageLabels";
-      const rangeTimeout = keymap.vimRangeTimeoutMs || 2000;
-
-      // Accumulate a count while a range operation is pending.
-      if (vimRange.current && chord >= "0" && chord <= "9") {
-        e.preventDefault();
-        const st = vimRange.current;
-        st.count = st.count * 10 + Number(chord);
-        if (st.timer) window.clearTimeout(st.timer);
-        st.timer = window.setTimeout(() => {
-          if (vimRange.current === st) {
-            runVimSingle(st.op, st.startId);
-            clearVimRange();
-          }
-        }, rangeTimeout);
-        setBulkProgress(`${st.key}${st.count}… (press ${st.key})`);
-        return;
-      }
-
-      const rop = rangeOps[chord];
-      if (rop) {
-        e.preventDefault();
-        // With an explicit bulk selection, act on it immediately (the TUI's "VIM
-        // BULK FIX") instead of starting a count sequence. A range means "the
-        // next N rows", which is meaningless once you've hand-picked a set — and
-        // the 2s buffer made the operation look hung (progress bar appears, sits
-        // idle, then the toast + removal land seconds later).
-        if (bulkMode && selected.size > 0) {
-          if (vimRange.current) clearVimRange();
-          runVimSingle(rop, selectedId ?? messages[0]?.id ?? "");
-          return;
-        }
-        const pending = vimRange.current;
-        if (pending && pending.key === chord) {
-          // Second press of the same key completes the range.
-          const count = pending.count || 1;
-          const startId = pending.startId;
-          clearVimRange();
-          runVimRange(rop, count, startId);
-          return;
-        }
-        // A different pending op: fire its single fallback now, then start fresh.
-        if (pending) {
-          const ps = pending;
-          clearVimRange();
-          runVimSingle(ps.op, ps.startId);
-        }
-        // Start a new sequence, capturing the row under the cursor as the target.
-        const startId = selectedId ?? messages[0]?.id ?? null;
-        if (!startId) return;
-        const st = { key: chord, op: rop, count: 0, startId, timer: 0 };
-        vimRange.current = st;
-        st.timer = window.setTimeout(() => {
-          if (vimRange.current === st) {
-            runVimSingle(rop, startId);
-            clearVimRange();
-          }
-        }, rangeTimeout);
-        setBulkProgress(`${chord}… (count then ${chord})`);
-        return;
-      }
-
-      // Any other key while a range is pending: fire the pending single on its
-      // original target, then fall through to handle this key normally.
-      if (vimRange.current) {
-        const ps = vimRange.current;
-        clearVimRange();
-        runVimSingle(ps.op, ps.startId);
-      }
-
-      // --- config-driven actions ---
-      const action = chordAction[chord];
-      if (!action) return;
-      e.preventDefault();
-      switch (action) {
-        case "summarize":
-          if (detail && aiEnabled && !bulkMode) void summarize(detail.id);
-          break;
-        case "regenerateSummary":
-          // Regenerate the active AI panel: the shown prompt, else the summary.
-          // (Uppercase of the summarize key — a real shortcut, e.g. Shift+Y.)
-          if (detail && !bulkMode) regenerateActive();
-          break;
-        case "prompt":
-          if (
-            aiPromptsEnabled &&
-            ((detail && !bulkMode) || (bulkMode && selected.size > 0))
-          )
-            setPromptsOpen(true);
-          break;
-        case "archive":
-          if (hasSel) void bulkAction("archive");
-          else if (detail) void doAction("archive", detail.id);
-          break;
-        case "quit":
-          void backend.Quit();
-          break;
-        case "trash":
-          if (hasSel) void bulkAction("trash");
-          else if (detail) void doAction("trash", detail.id);
-          break;
-        case "toggleRead":
-          if (hasSel) void bulkAction("unread");
-          else if (detail)
-            void doAction(detail.unread ? "read" : "unread", detail.id);
-          break;
-        case "manageLabels":
-          if (hasSel) setBulkLabels(true);
-          else if (detail) setLabelsFor(detail.id);
-          break;
-        case "compose":
-          setCompose({ mode: "new" });
-          break;
-        case "reply":
-          if (detail && !bulkMode) setCompose(replyInit(detail));
-          break;
-        case "forward":
-          if (detail && !bulkMode) setCompose(forwardInit(detail));
-          break;
-        case "search":
-          searchRef.current?.focus();
-          break;
-        case "refresh":
-          void load(activeQuery);
-          break;
-        case "loadMore":
-          void loadMore();
-          break;
-        case "drafts":
-          openDrafts();
-          break;
-        case "openGmail":
-          if (detail) openInGmail(detail.id);
-          break;
-        case "markdown":
-          if (detail && detail.html && detail.html.trim())
-            setViewHtml((v) => !v);
-          break;
-        case "bulkMode":
-          if (bulkMode) exitBulk();
-          else {
-            setBulkMode(true);
-            if (!selectedId && messages.length > 0)
-              setSelectedId(messages[0].id);
-          }
-          break;
-        case "bulkSelect":
-          if (bulkMode && selectedId) toggleSelect(selectedId);
-          break;
-        case "help":
-          setShowHelp(true);
-          break;
-        case "links":
-          if (detail) setLinksFor(detail.id);
-          break;
-        case "replyAll":
-          if (detail && !bulkMode) setCompose(replyAllInit(detail));
-          break;
-        case "saveMessage":
-          if (detail) saveMessage(detail.id);
-          break;
-        case "suggestLabel":
-          if (detail && aiEnabled && !bulkMode) void openSuggest(detail.id);
-          break;
-        case "obsidian":
-          if (detail && obsidianOn) sendObsidian(detail.id);
-          break;
-        case "slack":
-          if (detail && slackOn) forwardSlack(detail.id);
-          break;
-        case "commandMode":
-          setCmdOpen(true);
-          break;
-        case "threading":
-          if (detail && threadingOn) void toggleThread();
-          break;
-        case "savedQueries":
-          if (savedQueriesOn) void openQueries();
-          break;
-        case "saveQuery":
-          if (savedQueriesOn && activeQuery) setSaveQueryOpen(true);
-          break;
-        case "actionPlan":
-          if (actionPlanOn) void runActionPlan();
-          break;
-        case "themePicker":
-          if (themesOn) setThemePickerOpen(true);
-          break;
-        case "move":
-          if (hasSel) setBulkMove(true);
-          else if (detail && !bulkMode) setMoveFor(detail.id);
-          break;
-        case "toggleHeaders":
-          if (detail) {
-            setHeadersHidden(!headersHidden);
-            showToast(headersHidden ? "Headers visible" : "Headers hidden");
-          }
-          break;
-        case "searchFrom":
-          if (detail) quickSearch("from", detail);
-          break;
-        case "searchTo":
-          if (detail) quickSearch("to", detail);
-          break;
-        case "searchSubject":
-          if (detail) quickSearch("subject", detail);
-          break;
-        case "undo":
-          void runUndo();
-          break;
-        case "unread":
-          void load("is:unread");
-          break;
-        case "archived":
-          void load("in:archive");
-          break;
-        case "saveRaw":
-          if (detail) saveRawMessage(detail.id);
-          break;
-        case "attachments":
-          // Open the keyboard-navigable attachments picker (TUI's PickerAttachments).
-          // The reader still shows the same attachments as inline chips for the mouse.
-          if (detail && attachments.length > 0) setAttachmentsOpen(true);
-          break;
-        case "rsvp":
-          // Open the keyboard-navigable RSVP picker for the current invite
-          // (TUI's V). Only meaningful when the open message is an invite.
-          if (detail && invite?.isInvite) setRsvpPickerOpen(true);
-          break;
-      }
+    const ctx: KeydownCtx = {
+      attachmentsOpen, activeQuery, gPressedAt, vimRange, forwardSlack, themesOn,
+      accounts, accountsOpen, actionPlanOn, advOpen, aiEnabled, aiPromptsEnabled,
+      applyCategory, attachments, bulkAction, bulkLabels, bulkMode, bulkMove,
+      bulkPromptText, bumpZoom, chordAction, clearVimRange, cmdOpen, compose,
+      configOpen, csOpen, csQuery, detRulesOpen, detail, dismissAI,
+      doAction, draftsView, exitBulk, fullMessagesRef, headersHidden, invite,
+      keymap, labelsFor, linksFor, load, loadMore, localFilter,
+      messages, moveFor, obsidianOn, openDrafts, openInGmail, openMessage,
+      openQueries, openRules, openSuggest, plan, planActiveRef, planMove,
+      planNodesRef, planOpen, planPreview, previewMessage, promptManagerOpen, promptPreview,
+      promptsOpen, queriesOpen, quickSearch, readerBodyRef, readerFocused, regenerateActive,
+      resetZoom, rsvpPickerOpen, rulesEnabled, rulesOpen, runActionPlan, runUndo,
+      runVimRange, runVimSingle, saveMessage, saveQueryOpen, saveRawMessage, savedQueriesOn,
+      searchRef, selected, selectedId, sendObsidian, setAccountsOpen, setAdvOpen,
+      setAttachmentsOpen, setBulkLabels, setBulkMode, setBulkMove, setBulkProgress, setBulkPromptText,
+      setCmdOpen, setCompose, setConfigOpen, setCsIndex, setCsOpen, setCsQuery,
+      setDetail, setDraftsView, setExpandedCats, setHeadersHidden, setLabelsFor, setLinksFor,
+      setLocalFilter, setMessages, setMoveFor, setPlanExcluded, setPlanMove, setPlanOpen,
+      setPlanPreview, setPromptManagerOpen, setPromptPreview, setPromptsOpen, setQueriesOpen, setQuery,
+      setReaderFocused, setRsvpPickerOpen, setRulesOpen, setSaveQueryOpen, setSelected, setSelectedId,
+      setShowHelp, setStatsOpen, setSuggestFor, setThemePickerOpen, setViewHtml, showHelp,
+      showToast, slackOn, statsOpen, suggestFor, summarize, themePickerOpen,
+      threadingOn, toggleSelect, toggleThread, viewAnalyzerPrompt,
     };
+    const onKey = (e: KeyboardEvent) => handleKeyDown(e, ctx);
     window.addEventListener("keydown", onKey);
     return () => window.removeEventListener("keydown", onKey);
   }, [
