+++ desktop/frontend/src/keydownCtx.ts
@@ -1,0 +1,151 @@
+import type { Dispatch, MutableRefObject, RefObject, SetStateAction } from "react";
+import type {
+  KeyMap, AccountInfo, MessageSummary, MessageDetail, ActionPlanResult,
+  PlanCategory, Invite, Attachment,
+} from "./apiTypes";
+import type { ComposeInit } from "./Compose";
+import type { PlanNode } from "./planNodes";
+
+export type VimOp = "archive" | "trash" | "toggleRead" | "manageLabels";
+
+// Everything the keyboard handler (handleKeyDown/handleKeyMain) needs from App.
+// The keydown useEffect builds this each time its deps change; the ~700-line
+// handler then lives in keydownHandler.ts / keydownMain.ts instead of App.tsx.
+export interface KeydownCtx {
+  forwardSlack: (id: string) => void;
+  themesOn: boolean;
+  attachmentsOpen: boolean;
+  activeQuery: string;
+  gPressedAt: MutableRefObject<number>;
+  vimRange: MutableRefObject<{ key: string; op: VimOp; count: number; startId: string; timer: number } | null>;
+  accounts: AccountInfo[];
+  accountsOpen: boolean;
+  actionPlanOn: boolean;
+  advOpen: boolean;
+  aiEnabled: boolean;
+  aiPromptsEnabled: boolean;
+  applyCategory: (cat: PlanCategory, asMove?: boolean) => Promise<void>;
+  attachments: Attachment[];
+  bulkAction: (action: "archive" | "trash" | "read" | "unread") => Promise<void>;
+  bulkLabels: boolean;
+  bulkMode: boolean;
+  bulkMove: boolean;
+  bulkPromptText: string | null;
+  bumpZoom: (delta: number) => void;
+  chordAction: Record<string, string>;
+  clearVimRange: () => void;
+  cmdOpen: boolean;
+  compose: ComposeInit | null;
+  configOpen: boolean;
+  csOpen: boolean;
+  csQuery: string;
+  detRulesOpen: boolean;
+  detail: MessageDetail | null;
+  dismissAI: () => boolean;
+  doAction: (action: "archive" | "trash" | "read" | "unread", id: string) => Promise<void>;
+  draftsView: boolean;
+  exitBulk: () => void;
+  fullMessagesRef: MutableRefObject<MessageSummary[]>;
+  headersHidden: boolean;
+  invite: Invite | null;
+  keymap: KeyMap;
+  labelsFor: string | null;
+  linksFor: string | null;
+  load: (q: string) => Promise<void>;
+  loadMore: () => Promise<void>;
+  localFilter: boolean;
+  messages: MessageSummary[];
+  moveFor: string | null;
+  obsidianOn: boolean;
+  openDrafts: () => void;
+  openInGmail: (id: string) => void;
+  openMessage: (m: MessageSummary) => void;
+  openQueries: () => Promise<void>;
+  openRules: () => Promise<void>;
+  openSuggest: (id: string) => Promise<void>;
+  plan: ActionPlanResult | null;
+  planActiveRef: MutableRefObject<number>;
+  planMove: { kind: "email" | "category"; catIdx: number; id?: string } | null;
+  planNodesRef: MutableRefObject<PlanNode[]>;
+  planOpen: boolean;
+  planPreview: MessageDetail | null;
+  previewMessage: (m: MessageSummary) => void;
+  promptManagerOpen: boolean;
+  promptPreview: string | null;
+  promptsOpen: boolean;
+  queriesOpen: boolean;
+  quickSearch: (kind: "from" | "to" | "subject", d: MessageDetail) => void;
+  readerBodyRef: RefObject<HTMLDivElement>;
+  readerFocused: boolean;
+  regenerateActive: () => void;
+  resetZoom: () => void;
+  rsvpPickerOpen: boolean;
+  rulesEnabled: boolean;
+  rulesOpen: boolean;
+  runActionPlan: () => void;
+  runUndo: () => Promise<void>;
+  runVimRange: (op: VimOp, count: number, startId: string) => void;
+  runVimSingle: (op: VimOp, startId: string) => void;
+  saveMessage: (id: string) => void;
+  saveQueryOpen: boolean;
+  saveRawMessage: (id: string) => void;
+  savedQueriesOn: boolean;
+  searchRef: RefObject<HTMLInputElement>;
+  selected: Set<string>;
+  selectedId: string | null;
+  sendObsidian: (id: string) => void;
+  setAccountsOpen: Dispatch<SetStateAction<boolean>>;
+  setAdvOpen: Dispatch<SetStateAction<boolean>>;
+  setAttachmentsOpen: Dispatch<SetStateAction<boolean>>;
+  setBulkLabels: Dispatch<SetStateAction<boolean>>;
+  setBulkMode: Dispatch<SetStateAction<boolean>>;
+  setBulkMove: Dispatch<SetStateAction<boolean>>;
+  setBulkProgress: Dispatch<SetStateAction<string>>;
+  setBulkPromptText: Dispatch<SetStateAction<string | null>>;
+  setCmdOpen: Dispatch<SetStateAction<boolean>>;
+  setCompose: Dispatch<SetStateAction<ComposeInit | null>>;
+  setConfigOpen: Dispatch<SetStateAction<boolean>>;
+  setCsIndex: Dispatch<SetStateAction<number>>;
+  setCsOpen: Dispatch<SetStateAction<boolean>>;
+  setCsQuery: Dispatch<SetStateAction<string>>;
+  setDetail: Dispatch<SetStateAction<MessageDetail | null>>;
+  setDraftsView: Dispatch<SetStateAction<boolean>>;
+  setExpandedCats: Dispatch<SetStateAction<Set<string>>>;
+  setHeadersHidden: Dispatch<SetStateAction<boolean>>;
+  setLabelsFor: Dispatch<SetStateAction<string | null>>;
+  setLinksFor: Dispatch<SetStateAction<string | null>>;
+  setLocalFilter: Dispatch<SetStateAction<boolean>>;
+  setMessages: Dispatch<SetStateAction<MessageSummary[]>>;
+  setMoveFor: Dispatch<SetStateAction<string | null>>;
+  setPlanExcluded: Dispatch<SetStateAction<Set<string>>>;
+  setPlanMove: Dispatch<SetStateAction<{ kind: "email" | "category"; catIdx: number; id?: string } | null>>;
+  setPlanOpen: Dispatch<SetStateAction<boolean>>;
+  setPlanPreview: Dispatch<SetStateAction<MessageDetail | null>>;
+  setPromptManagerOpen: Dispatch<SetStateAction<boolean>>;
+  setPromptPreview: Dispatch<SetStateAction<string | null>>;
+  setPromptsOpen: Dispatch<SetStateAction<boolean>>;
+  setQueriesOpen: Dispatch<SetStateAction<boolean>>;
+  setQuery: Dispatch<SetStateAction<string>>;
+  setReaderFocused: Dispatch<SetStateAction<boolean>>;
+  setRsvpPickerOpen: Dispatch<SetStateAction<boolean>>;
+  setRulesOpen: Dispatch<SetStateAction<boolean>>;
+  setSaveQueryOpen: Dispatch<SetStateAction<boolean>>;
+  setSelected: Dispatch<SetStateAction<Set<string>>>;
+  setSelectedId: Dispatch<SetStateAction<string | null>>;
+  setShowHelp: Dispatch<SetStateAction<boolean>>;
+  setStatsOpen: Dispatch<SetStateAction<boolean>>;
+  setSuggestFor: Dispatch<SetStateAction<string | null>>;
+  setThemePickerOpen: (v: boolean) => void;
+  setViewHtml: Dispatch<SetStateAction<boolean>>;
+  showHelp: boolean;
+  showToast: (m: string) => void;
+  slackOn: boolean;
+  statsOpen: boolean;
+  suggestFor: string | null;
+  summarize: (id: string, force?: boolean) => Promise<void>;
+  themePickerOpen: boolean;
+  threadingOn: boolean;
+  toggleSelect: (id: string) => void;
+  toggleThread: () => Promise<void>;
+  viewAnalyzerPrompt: () => Promise<void>;
+}
